Improved UI + Input Validation
More descriptive UI 
All API calls are now validated to make sure they are in correct format

+++ server.py
@@ -5,9 +5,7 @@ import requests
 
 SteamKey = '21C5CFC0D481BE7C44CEBCAFCE909614'  # Pi SteamKey
 PortNumber = 5011  # Pi port number
-#HostURL = 'http://cs2s.yorkdc.net'
 #HostURL = 'cs2s.yorkdc.net'
-#HostURL = '192.168.0.32'
 SteamID = '76561198058093131'
 
 restServer = Flask(__name__)
@@ -37,24 +35,48 @@ def after_request(response):
 def test():
     return "Hello"
 
-#Takes data from address and inserts it into CoD table
+
+#Takes data from API call and inserts it into CoD table
 @restServer.route("/sendCoD/<playerID>/<mapID>/<gameModeID>/<kills>/<deaths>/<assists>/<accuracy>", methods=['GET'])
 def sendCoD(playerID, mapID, gameModeID, kills, deaths, assists, accuracy):
-    cursor.execute("INSERT INTO CoD (playerID, mapID, gameModeID, kills, deaths, assists, accuracy) VALUES (%s, %s, %s, %s, %s, %s, %s)", (playerID, mapID, gameModeID, kills, deaths, assists, accuracy))
-    conn.commit()
-    return "Success"
+    if(int(playerID) > 0):
+        if (mapID != "Containers" and mapID != "Village" and mapID != "Ocean Vila"):
+            return "Invalid Map"
+        elif (gameModeID != "Team Death Match" and gameModeID !="Free For All" and gameModeID !="Infected"):
+            return "Invalid Gamemode"
+        elif (int(kills) > 100 or int(kills) < 0):
+            return "Kills value is impossible"
+        elif (int(deaths) > 100 or int(deaths) < 0):
+            return "Deaths value is impossible"
+        elif (int(assists) > 100 or int(assists) < 0):
+            return "Assists value is impossible"
+        elif (int(accuracy) > 100 or int(accuracy) < 0):
+            return "Accuracy value is impossible"
+        else:
+            cursor.execute("INSERT INTO CoD (playerID, mapID, gameModeID, kills, deaths, assists, accuracy) VALUES (%s, %s, %s, %s, %s, %s, %s)", (playerID, mapID, gameModeID, kills, deaths, assists, accuracy))
+            conn.commit()
+            return "Success"
+    else:
+        return "Invalid playerID"
 
 #Delete Entry From CoD Table providing authcode matches
 @restServer.route("/delCoD/<EntryID>/<AuthCode>", methods=['POST'])
 def delCoD(EntryID, AuthCode):
-    if AuthCode == "123":#Check the authcode matches
-        intEntryID = int(EntryID)#Converts data to a int
-        #cursor.execute("DELETE FROM D2 (EntryID) VALUES (%d)", (intEntryID))
-        cursor.execute("DELETE FROM CoD WHERE EntryID=" + EntryID)#Executes the command to delete Entry from table
-        conn.commit()
-        return "Success"
+    if(int(EntryID) > 0):
+        if AuthCode == "123":#Check the authcode matches
+            intEntryID = int(EntryID)#Converts data to a int
+            rows_count = cursor.execute("SELECT * FROM CoD WHERE entryID='" + EntryID + "'")
+            if rows_count > 0:
+                cursor.execute("DELETE FROM CoD WHERE EntryID=" + EntryID)#Executes the command to delete Entry from table
+                conn.commit()
+                return "Success"
+            else:
+                return "Entry doesn't exist"
+        else:
+            return "Incorrect AuthCode"
     else:
-        return "Incorrect AuthCode"
+        return "Invalid EntryID"
+
 
 #Returns all data in CoD table in JSON format
 @restServer.route("/getCoD", methods=['GET'])
@@ -64,58 +86,91 @@ def getCoD():
     cursor.execute("SELECT * FROM CoD")#Gets all rows from CoD
     result = cursor.fetchall()
     for i in result:#for each result
-    #put it in JSON format
+    #Creates Key Value pairs
         statList = {'entryID': i[0], 'playerID': i[1], 'mapID': i[2], 'gameModeID': i[3], 'kills': i[4], 'deaths': i[5], 'assists': i[6], 'accuracy': i[7]}
-        resultList.append(statList) #add it to list
+        resultList.append(statList) #adds it to list
     records = {'results': resultList}
-    return jsonify(records) #return list of records
+    return jsonify(records) #return list of records in JSON format
 
 #Retuns specified player data from CoD table in JSON format
 @restServer.route("/getCoD/<playerID>", methods=['GET'])
 def getCoDWithID(playerID):
-    resultList = []
-    records = []
-    cursor.execute("SELECT * FROM CoD WHERE playerID='" + playerID + "'")
-    result = cursor.fetchall()
-    for i in result:
-        statList = {'entryID': i[0], 'playerID': i[1], 'mapID': i[2], 'gameModeID': i[3], 'kills': i[4], 'deaths': i[5], 'assists': i[6], 'accuracy': i[7]}
-        resultList.append(statList)
-    records = {'results': resultList}
-    return jsonify(records)
+    if(int(playerID) > 0):
+        resultList = []
+        records = []
+        cursor.execute("SELECT * FROM CoD WHERE playerID='" + playerID + "'")
+        result = cursor.fetchall()
+        for i in result:
+            #Creates Key Value pairs
+            statList = {'entryID': i[0], 'playerID': i[1], 'mapID': i[2], 'gameModeID': i[3], 'kills': i[4], 'deaths': i[5], 'assists': i[6], 'accuracy': i[7]}
+            resultList.append(statList)
+            records = {'results': resultList}
+            return jsonify(records)
+    else:
+        return "Invalid PlayerID"
 
 #Retuns specified player where map is specified data from CoD table in JSON format
 @restServer.route("/getCoD/<playerID>/<gameMode>", methods=['GET'])
 def getCoDWithIDMap(playerID, gameMode):
-    resultList = []
-    records = []
-    cursor.execute("SELECT * FROM CoD WHERE playerID='" + playerID + "' AND gameModeID='" + gameMode +"'")
-    #cursor.execute("SELECT * FROM CoD WHERE mapID='Village' AND playerID='13'")
-    result = cursor.fetchall()
-    for i in result:
-        statList = {'entryID': i[0], 'playerID': i[1], 'mapID': i[2], 'gameModeID': i[3], 'kills': i[4], 'deaths': i[5], 'assists': i[6], 'accuracy': i[7]}
-        resultList.append(statList)
-    records = {'results': resultList}
-    return jsonify(records)
+    if(int(playerID) > 0):
+        if (gameMode != "Team Death Match" and gameMode !="Free For All" and gameMode !="Infected"):
+            return "Invalid Gamemode"
+        else:
+            resultList = []
+            records = []
+            cursor.execute("SELECT * FROM CoD WHERE playerID='" + playerID + "' AND gameModeID='" + gameMode +"'")
+            #cursor.execute("SELECT * FROM CoD WHERE mapID='Village' AND playerID='13'")
+            result = cursor.fetchall()
+            for i in result:
+                #Creates Key Value pairs
+                statList = {'entryID': i[0], 'playerID': i[1], 'mapID': i[2], 'gameModeID': i[3], 'kills': i[4], 'deaths': i[5], 'assists': i[6], 'accuracy': i[7]}
+                resultList.append(statList)
+                records = {'results': resultList}
+                return jsonify(records)
+    else:
+        return "Invalid PlayerID"
 
 
-#Send Destiny 2 data to MySQL DB
+#Takes data from API call and inserts in into Destiny 2 table to MySQL DB
 @restServer.route("/sendD2/<playerID>/<mapID>/<MotesDeposited>/<HostilesDefeated>/<GuardiansDefeated>/<MotesLost>/<PrimevalHealed>", methods=['GET'])
 def sendD2(playerID, mapID, MotesDeposited, HostilesDefeated, GuardiansDefeated, MotesLost, PrimevalHealed):
-    cursor.execute("INSERT INTO D2 (playerID, mapID, MotesDeposited, HostilesDefeated, GuardiansDefeated, MotesLost, PrimevalHealed) VALUES (%s, %s, %s, %s, %s, %s, %s)", (playerID, mapID, MotesDeposited, HostilesDefeated, GuardiansDefeated, MotesLost, PrimevalHealed))
-    conn.commit()
-    return "Success"
+    if(int(playerID) > 0):
+        if (mapID != "Kell's Grave" and mapID != "Legion's Folly" and mapID != "Catherdral of Scars" and mapID != "Emerald Coast"):
+            return "Invalid Map"
+        elif (int(MotesDeposited) > 75 or int(MotesDeposited) < 0):
+            return "Motes Bank value is impossible"
+        elif (int(HostilesDefeated) > 500 or int(HostilesDefeated) < 0):
+            return "Hostile Kills value is impossible"
+        elif (int(GuardiansDefeated) > 100 or int(GuardiansDefeated) < 0):
+            return "Assists value is impossible"
+        elif (int(MotesLost) > 500 or int(MotesLost) < 0):
+            return "Accuracy value is impossible"
+        elif (int(PrimevalHealed) > 100 or int(PrimevalHealed) < 0):
+            return "Accuracy value is impossible"
+        else:
+            cursor.execute("INSERT INTO D2 (playerID, mapID, MotesDeposited, HostilesDefeated, GuardiansDefeated, MotesLost, PrimevalHealed) VALUES (%s, %s, %s, %s, %s, %s, %s)", (playerID, mapID, MotesDeposited, HostilesDefeated, GuardiansDefeated, MotesLost, PrimevalHealed))
+            conn.commit()
+            return "Success"
+    else:
+        return "Invalid PlayerID"
 
 #Deletes a entry from D2 table providing authcode matches
 @restServer.route("/delD2/<EntryID>/<AuthCode>", methods=['POST'])
 def delD2(EntryID, AuthCode):
-    if AuthCode == "123":
-        intEntryID = int(EntryID)
-        #cursor.execute("DELETE FROM D2 (EntryID) VALUES (%d)", (intEntryID))
-        cursor.execute("DELETE FROM D2 WHERE EntryID=" + EntryID)
-        conn.commit()
-        return "Success"
+    if(int(EntryID) > 0):
+        if AuthCode == "123":#Check the authcode matches
+            intEntryID = int(EntryID)#Converts data to a int
+            rows_count = cursor.execute("SELECT * FROM D2 WHERE entryID='" + EntryID + "'")
+            if rows_count > 0:
+                cursor.execute("DELETE FROM D2 WHERE EntryID=" + EntryID)#Executes the command to delete Entry from table
+                conn.commit()
+                return "Success"
+            else:
+                return "Entry doesn't exist"
+        else:
+            return "Incorrect AuthCode"
     else:
-        return "Incorrect AuthCode"
+        return "Invalid EntryID"
 
 #Returns all data in D2 table in JSON format
 @restServer.route("/getD2", methods=['GET'])
@@ -125,64 +180,82 @@ def getD2():
     cursor.execute("SELECT * FROM D2")
     result = cursor.fetchall()
     for i in result:
+        #Creates Key Value pairs
         statList = {'EntryID': i[0], 'playerID': i[1], 'mapID': i[2], 'MotesDeposited': i[3], 'HostilesDefeated': i[4], 'GuardiansDefeated': i[5], 'MotesLost': i[6], 'PrimevalHealed': i[7]}
         resultList.append(statList)
     records = {'results': resultList}
     return jsonify(records)
 
-#Retuns specified player data from CoD table in JSON format
+#Retuns specified player data from D2 table in JSON format
 @restServer.route("/getD2/<playerID>", methods=['GET'])
 def getD2WithID(playerID):
-    resultList = []
-    records = []
-    cursor.execute("SELECT * FROM D2 WHERE playerID='" + playerID + "'")
-    result = cursor.fetchall()
-    for i in result:
-        statList = {'EntryID': i[0], 'playerID': i[1], 'mapID': i[2], 'MotesDeposited': i[3], 'HostilesDefeated': i[4], 'GuardiansDefeated': i[5], 'MotesLost': i[6], 'PrimevalHealed': i[7]}
-        resultList.append(statList)
-    records = {'results': resultList}
-    return jsonify(records)
+    if (int(playerID) > 0):
+        resultList = []
+        records = []
+        cursor.execute("SELECT * FROM D2 WHERE playerID='" + playerID + "'")
+        result = cursor.fetchall()
+        for i in result:
+            #Creates Key Value pairs
+            statList = {'EntryID': i[0], 'playerID': i[1], 'mapID': i[2], 'MotesDeposited': i[3], 'HostilesDefeated': i[4], 'GuardiansDefeated': i[5], 'MotesLost': i[6], 'PrimevalHealed': i[7]}
+            resultList.append(statList)
+            records = {'results': resultList}
+            return jsonify(records)
+    else:
+        return "Invalid Player ID"
 
 
-#Call with a steam player id and it will return all their game using the Steam API
+#Uses Steam ID from API call to call Steam API to retrieve list of games associated with that account
 @restServer.route("/getSteamGames/<steamID>", methods=['GET'])
 def getSteamGames(steamID):
-    api_url = 'http://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/?key=' + SteamKey + '&steamid=' + steamID +'&format=json'
-    response = requests.get(api_url)
-    if response.status_code == 200:
-        return (response.content)
+    if (int(steamID) <= 99999999999999999 and int(steamID) >= 10000000000000000):#Valid Steam IDs are between these numbers
+        api_url = 'http://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/?key=' + SteamKey + '&steamid=' + steamID +'&format=json'
+        response = requests.get(api_url)
+        if response.status_code == 200:
+            #Returns list of games in JSON format
+            return (response.content)
+        else:
+            return "Invalid Steam ID"
     else:
-        return "Fail"
+        return "Steam ID is incorrect length"
 
-#Call with id of a steam game and it will return all details about it using Steam API
+#Uses Steam game ID to call Steam API to retrieve a description of that game ID
 @restServer.route("/getSteamGameDes/<gameID>", methods=['GET'])
 def getSteamGameDes(gameID):
-    api_url = 'http://api.steampowered.com/ISteamUserStats/GetSchemaForGame/v2/?key=' + SteamKey + '&appid=' + gameID
-    response = requests.get(api_url)
-    if response.status_code == 200:
-        return (response.content)
+    if (int(gameID) > 0):#Valid Steam game IDs are all positive
+        api_url = 'http://api.steampowered.com/ISteamUserStats/GetSchemaForGame/v2/?key=' + SteamKey + '&appid=' + gameID
+        response = requests.get(api_url)
+        if response.status_code == 200:
+            return (response.content)
+        else:
+            return "Invalid Steam Game ID"
     else:
         return "Fail"
 
-#Call with id of a steam game and it will return all details about it using Steam API
+#Uses Steam Player ID to call Steam API to retrieve a list of Steam ID that are friends with the provided Steam ID
 @restServer.route("/getSteamFriends/<steamID>", methods=['GET'])
 def getSteamFriends(steamID):
-    api_url = 'http://api.steampowered.com/ISteamUser/GetFriendList/v0001/?key=' + SteamKey + '&steamid=' + steamID + '&relationship=friend'
-    response = requests.get(api_url)
-    if response.status_code == 200:
-        return (response.content)
+    if (int(steamID) <= 99999999999999999 and int(steamID) >= 10000000000000000):#Valid Steam IDs are between these numbers
+        api_url = 'http://api.steampowered.com/ISteamUser/GetFriendList/v0001/?key=' + SteamKey + '&steamid=' + steamID + '&relationship=friend'
+        response = requests.get(api_url)
+        if response.status_code == 200:
+            return (response.content)
+        else:
+            return "Invalid Steam ID"
     else:
-        return "Fail"
+        return "Steam ID is incorrect length"
 
-#Call with id of a steam game and it will return all details about it using Steam API
+#Uses Steam ID to call Steam API to retrieve a JSON object of that IDs Steam profile
 @restServer.route("/getSteamProfile/<steamID>", methods=['GET'])
 def getSteamProfile(steamID):
-    api_url = 'http://api.steampowered.com/ISteamUser/GetPlayerSummaries/v0002/?key=' + SteamKey + '&steamids=' + steamID
-    response = requests.get(api_url)
-    if response.status_code == 200:
-        return (response.content)
+    if (int(steamID) <= 99999999999999999 and int(steamID) >= 10000000000000000):#Valid Steam IDs are between these numbers
+        api_url = 'http://api.steampowered.com/ISteamUser/GetPlayerSummaries/v0002/?key=' + SteamKey + '&steamids=' + steamID
+        response = requests.get(api_url)
+        if response.status_code == 200:
+            return (response.content)
+        else:
+            return "Invalid Steam ID"
     else:
-        return "Fail"
+        return "Steam ID is incorrect length"
